# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentación completa de la API del sistema",
        "contact": {
            "email": "info@appnet.dev"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Obtener lista de contactos",
                "description": "Devuelve todos los contactos del usuario autenticado",
                "operationId": "contactsIndex",
                "responses": {
                    "200": {
                        "description": "Lista de contactos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Nombre del contacto"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+34600000000"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Dirección del contacto"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "contacto@ejemplo.com"
                                                    },
                                                    "web": {
                                                        "type": "string",
                                                        "example": "https://ejemplo.com"
                                                    },
                                                    "telegram": {
                                                        "type": "string",
                                                        "example": "@usuario_telegram"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Crear un nuevo contacto",
                "description": "Crea un nuevo contacto para el usuario autenticado",
                "operationId": "contactsStore",
                "requestBody": {
                    "description": "Datos del nuevo contacto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre del contacto"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+34600000000"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Dirección del contacto"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contacto@ejemplo.com"
                                    },
                                    "web": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://ejemplo.com"
                                    },
                                    "telegram": {
                                        "type": "string",
                                        "example": "@usuario_telegram"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nombre del contacto"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+34600000000"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Dirección del contacto"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "contacto@ejemplo.com"
                                                },
                                                "web": {
                                                    "type": "string",
                                                    "example": "https://ejemplo.com"
                                                },
                                                "telegram": {
                                                    "type": "string",
                                                    "example": "@usuario_telegram"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contacto creado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "whatsappToken": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "scrapingToken": {
                "type": "apiKey",
                "description": "Token de autenticación para el servicio de scraping",
                "name": "token",
                "in": "query"
            }
        }
    },
...
}